refactor(e2e-suite): split off assets and receiveAccount from tradingPage
