+++ suite/e2e/tests/trading/buy-negative.test.ts
@@ -15,7 +15,7 @@ test.describe('Trading - Buy Negative scenarios', { tag: ['@webOnly', '@T3W1', '
             await walletPage.openTradingGlobalButton.click();
 
             await expect(tradingPage.youPayFiatInput).toHaveValue(''); // waits for trading form to load
-            await tradingPage.buyAssetPickerInput.click({ trial: true }); // checking actionability of the dropdown, which means page is properly loaded
+            await tradingPage.assets.buyAssetPickerInput.click({ trial: true }); // checking actionability of the dropdown, which means page is properly loaded
 
             await tradingPage.selectFiatCurrency('eur');
         });
