+++ suite/e2e/tests/trading/swap-coin-to-token.test.ts
@@ -61,7 +61,7 @@ test.describe('Trading - Swap coin to token', { tag: ['@webOnly', '@T3W1', '@T3T
                 },
                 receiveAddress,
                 selectReceiveAddress: async () => {
-                    await tradingPage.selectSuiteReceiveAccount(0);
+                    await tradingPage.receiveAccount.selectSuiteReceiveAccount(0);
                 },
             });
         });
